feat: rename swiftErrorArchiver content
