+++ SwiftEventShooter/Sources/SwiftEventShooter/EventController/EventStreamControllerInterface.swift
@@ -1,13 +1,13 @@
 //
-//  EventControllerInterface.swift
-//  SwiftErrorArchiver
+//  EventStreamControllerInterface.swift
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/20/24.
 //
 
 import Foundation
 
-public protocol EventControllerInterface: Sendable {
+public protocol EventStreamControllerInterface: Sendable {
   func post(_ event: some EventInterface) async
   func sendPendingEvents() async
   func configure()
