+++ SwiftEventShooter/Sources/SwiftEventShooter/EventStorage/Commons/EventWithDate.swift
@@ -1,6 +1,6 @@
 //
 //  EventWithDate.swift
-//  SwiftErrorArchiver
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/20/24.
 //
