+++ SwiftEventShooter/Sources/SwiftEventShooter/EventStorage/EventStorageControllerInterface.swift
@@ -1,6 +1,6 @@
 //
 //  EventStorageControllerInterface.swift
-//  SwiftErrorArchiver
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/20/24.
 //
