+++ SwiftEventShooter/Sources/SwiftEventShooter/NetworkMonitor/MockNetworkMonitor.swift
@@ -1,6 +1,6 @@
 //
-//  File.swift
-//  SwiftErrorArchiver
+//  MockNetworkMonitor.swift
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/28/24.
 //
