+++ SwiftEventShooter/Sources/SwiftEventShooter/NetworkMonitor/NetworkMonitor.swift
@@ -1,6 +1,6 @@
 //
 //  NetworkMonitor.swift
-//  SwiftErrorArchiver
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/23/24.
 //
