+++ SwiftEventShooter/Sources/SwiftEventShooter/PendingStreamManager/TimerManager.swift
@@ -1,6 +1,6 @@
 //
-//  File.swift
-//  SwiftErrorArchiver
+//  TimerManager.swift
+//  SwiftEventShooter
 //
 //  Created by MaraMincho on 12/28/24.
 //
@@ -8,22 +8,22 @@
 import Foundation
 
 final class TimerManager: @unchecked Sendable {
-    private var timer: DispatchSourceTimer?
-    private let interval: TimeInterval
+  private var timer: DispatchSourceTimer?
+  private let interval: TimeInterval
 
-    init(interval: TimeInterval) {
-        self.interval = interval
-    }
+  init(interval: TimeInterval) {
+    self.interval = interval
+  }
 
-    func startTimer(action: @Sendable @escaping () -> Void) {
-        timer = DispatchSource.makeTimerSource(queue: .global())
-        timer?.schedule(deadline: .now(), repeating: interval)
-        timer?.setEventHandler(handler: action)
-        timer?.resume()
-    }
+  func startTimer(action: @Sendable @escaping () -> Void) {
+    timer = DispatchSource.makeTimerSource(queue: .global())
+    timer?.schedule(deadline: .now(), repeating: interval)
+    timer?.setEventHandler(handler: action)
+    timer?.resume()
+  }
 
-    func stopTimer() {
-        timer?.cancel()
-        timer = nil
-    }
+  func stopTimer() {
+    timer?.cancel()
+    timer = nil
+  }
 }
