+++ SwiftEventShooterExample/SwiftEventShooterExample/SwiftEventShotterExample.swift
@@ -1,5 +1,5 @@
 //
-//  SwiftErrorArchiverExampleApp.swift
+//  SwiftEventShotterExample.swift
 //  SwiftErrorArchiverExample
 //
 //  Created by MaraMincho on 12/23/24.
@@ -9,9 +9,9 @@ import SwiftUI
 
 @main
 struct SwiftErrorArchiverExampleApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
+  var body: some Scene {
+    WindowGroup {
+      ContentView()
     }
+  }
 }
