+++ SwiftEventShooterExample/SwiftEventShooterTests/SwiftErrorArchiverExampleTests.swift
@@ -5,13 +5,11 @@
 //  Created by MaraMincho on 12/23/24.
 //
 
-import Testing
 @testable import SwiftErrorArchiverExample
+import Testing
 
 struct SwiftErrorArchiverExampleTests {
-
-    @Test func example() async throws {
-        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
-    }
-
+  @Test func example() async throws {
+    // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+  }
 }
